% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_weight.R
\name{calculate_weight}
\alias{calculate_weight}
\title{Calculate the survey weights for the observations}
\usage{
calculate_weight(observation)
}
\arguments{
\item{observation}{the dataframe with observations. Should be the output of \code{\link{read_observation}}}
}
\description{
Adds the survey weights and the stratum to the observations
}

