% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_attribute}
\alias{connect_attribute}
\title{Make a git connection to the attribute}
\usage{
connect_attribute(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Make a git connection to the attribute
}

