% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_raw}
\alias{connect_raw}
\title{Make a git connection to the raw data}
\usage{
connect_raw(result.channel, username = character(0),
  password = character(0))
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}

\item{username}{The username for a https connection. Leave default for a ssh
connection.}

\item{password}{The password for the connection.}
}
\description{
Make a git connection to the raw data
}

