% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect_rawdata.R
\name{connect_rawdata}
\alias{connect_rawdata}
\title{Connect to the rawdata}
\usage{
connect_rawdata(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Connect to the rawdata
}

