% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_source}
\alias{connect_source}
\title{Connect to the source data}
\usage{
connect_source(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Connect to the source data
}

