% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_manifest_objects.R
\name{copy_manifest_objects}
\alias{copy_manifest_objects}
\title{Copy all object of the most recent manifest}
\usage{
copy_manifest_objects(base, project, target_folder = ".", overwrite = FALSE)
}
\description{
Copy all object of the most recent manifest
}
