% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id}
\alias{datasource_id}
\title{Get the id of the data source}
\usage{
datasource_id(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Get the id of the data source
}

