% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_field_id.R
\name{get_field_id}
\alias{get_field_id}
\title{Retrieve the id for a field in a table}
\usage{
get_field_id(repo, table_name, field_name)
}
\arguments{
\item{repo}{a [git2rdata::repository()] object}

\item{table_name}{The name of the table.}

\item{field_name}{The name of the field.}
}
\description{
Retrieve the id for a field in a table
}
