% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_lincomb.R
\name{get_linear_lincomb}
\alias{get_linear_lincomb}
\title{Calculate the matrix for linear combinations of strata}
\usage{
get_linear_lincomb(
  dataset,
  time.var,
  stratum.var = "stratum",
  stratum_weights,
  formula
)
}
\arguments{
\item{dataset}{the raw dataset}

\item{time.var}{the name of the time variable}

\item{stratum.var}{the name of the stratum variable}

\item{stratum_weights}{a dataframe with stratum weights}

\item{formula}{the formula for the trend component}
}
\description{
Calculate the matrix for linear combinations of strata
}
