% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonlinear_lincomb.R
\name{get_nonlinear_lincomb}
\alias{get_nonlinear_lincomb}
\title{Calculate the matrix for linear combinations of strata}
\usage{
get_nonlinear_lincomb(
  dataset,
  stratum_weights,
  time.var,
  label.var,
  stratum.var = "stratum"
)
}
\arguments{
\item{dataset}{the raw dataset}

\item{stratum_weights}{a dataframe with stratum weights}

\item{time.var}{the name of the time variable}

\item{label.var}{an optional variable used to label the time variable, `time.var` will be used when missing}

\item{stratum.var}{the name of the stratum variable}
}
\description{
Calculate the matrix for linear combinations of strata
}
