% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_trend.R
\name{moving_trend}
\alias{moving_trend}
\title{Create weight matrix for linear trend over a moving window}
\usage{
moving_trend(n_total = 20, n_window = 10, start = 0)
}
\arguments{
\item{n_total}{Length of the time series.
Defaults to \code{20}.}

\item{n_window}{Size of the moving window.
Defaults to \code{10}.}

\item{start}{Optional offset for the time series.
Defaults to \code{0}.}
}
\description{
Create weight matrix for linear trend over a moving window
}
