% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_comparison.R
\name{prepare_analysis_comparison}
\alias{prepare_analysis_comparison}
\title{Prepare comparison among models}
\usage{
prepare_analysis_comparison(frequency, species_group, location_group,
  models, base, project, overwrite = FALSE)
}
\arguments{
\item{frequency}{a indicator of the frequency of the analysis}

\item{species_group}{the species_group ID}

\item{location_group}{the location_group ID}

\item{models}{a data.frame with the metadata of the models}

\item{base}{the base location to store the model}

\item{project}{will be a relative path within the base location}

\item{overwrite}{should an existing object be overwritten? Defaults to TRUE}
}
\description{
Prepare comparison among models
}
