% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_analysis_lrt.R
\name{prepare_analysis_lrt}
\alias{prepare_analysis_lrt}
\title{Prepare likelihood ratio tests among models}
\usage{
prepare_analysis_lrt(dataset, raw.connection, analysis.path)
}
\arguments{
\item{dataset}{A data.frame with the parent analyses}

\item{raw.connection}{a git-connection object to write the output to}

\item{analysis.path}{the path to store the rda files for the analysis}
}
\description{
Prepare likelihood ratio tests among models
}

