% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare all datasets}
\usage{
prepare_dataset(result, origin, repo, scheme.id, end.date = Sys.time(),
  verbose = TRUE, push = FALSE, ...)
}
\arguments{
\item{result}{An open ODBC connection to the results database}

\item{origin}{`dbplyr::src_dbi` connection to the source database}

\item{repo}{a \code{\link[git2rdata]{repository}} object}

\item{scheme.id}{the fingerprint of the scheme}

\item{end.date}{the latest date to import. Default to now.}

\item{verbose}{display the progress}

\item{push}{push the changes to the repository. Defaults to `FALSE`.}

\item{...}{arguments passed to \code{\link[git2rdata]{commit}} and
\code{\link[git2rdata]{push}}.}
}
\description{
Prepare all datasets and store the version into a git repository
}
