% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_location.R
\name{prepare_dataset_location}
\alias{prepare_dataset_location}
\title{import and store location metadata}
\usage{
prepare_dataset_location(origin, result, repo, scheme.id, end.date)
}
\arguments{
\item{origin}{`dbplyr::src_dbi` connection to the source database}

\item{result}{An open ODBC connection to the results database}

\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{scheme.id}{the fingerprint of the scheme}

\item{end.date}{the latest date to import. Default to now.}
}
\description{
import and store location metadata
}
