% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations and save them to git and the results database}
\usage{
prepare_dataset_observation(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Read the observations and save them to git and the results database
}

