% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species groups and save them to git and the results database}
\usage{
prepare_dataset_species(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Read the species groups and save them to git and the results database
}

