% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset_observation_species.R
\name{prepare_dataset_species_observation}
\alias{prepare_dataset_species_observation}
\title{Read the counts and save them to git}
\usage{
prepare_dataset_species_observation(this.species, observation, develop = TRUE)
}
\arguments{
\item{this.species}{a dataframe with the ExternalCode and the SpeciesGroupID}

\item{observation}{a data.frame with observations}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Read the counts and save them to git
}

