% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_observation.R
\name{read_observation}
\alias{read_observation}
\title{Return the set of all visited locations}
\usage{
read_observation(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Return the set of all visited locations
}

