% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_observation_species.R
\name{read_observation_species}
\alias{read_observation_species}
\title{Read the observations of a given species}
\usage{
read_observation_species(species.id, develop = TRUE)
}
\arguments{
\item{species.id}{The external id of the species}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Read the observations of a given species
}

