% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_specieslist.R
\name{read_specieslist}
\alias{read_specieslist}
\title{Read the list of species}
\usage{
read_specieslist(source.channel)
}
\arguments{
\item{source.channel}{An open ODBC channel to the source database}
}
\value{
A list with two components: \code{Species} holds the names of all species, \code{Speciesgroup} lists the species per user defined group.
}
\description{
Read the list of species
}
\examples{
read_specieslist()
}

