% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{result_datasource_id}
\alias{result_datasource_id}
\title{Get the id of the data source of the results}
\usage{
result_datasource_id(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Get the id of the data source of the results
}

