% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_results.R
\name{retrieve_results}
\alias{retrieve_results}
\title{Retrieve the relevant results}
\usage{
retrieve_results(
  base,
  project,
  source_repo,
  target_repo,
  verbose = TRUE,
  strict = TRUE
)
}
\arguments{
\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{source_repo}{A \code{\link[git2rdata:repository]{git2rdata::repository()}} object containing the raw data}

\item{target_repo}{A \code{\link[git2rdata:repository]{git2rdata::repository()}} object containing the package
code.}

\item{verbose}{display the progress}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}
}
\description{
Retrieve the relevant results
}
