% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scheme_id.R
\name{scheme_id}
\alias{scheme_id}
\title{Get the id of the scheme}
\usage{
scheme_id(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Get the id of the scheme
}

