% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_relevant.R
\name{select_relevant}
\alias{select_relevant}
\title{Select the relevant observation of a species}
\usage{
select_relevant(observation, observation.species)
}
\arguments{
\item{observation}{the dataframe with observations. Should be the output of
\code{\link{read_observation}}}

\item{observation.species}{The output from \code{\link{read_observation_species}}}
}
\description{
Relevant locations have at least two observations in different years. Relevant periods have average numbers of at least 5% of the most important period.
}
