% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage}
\alias{storage}
\title{store the n2kModel and retrieve its fingerprints}
\usage{
storage(model, base, project, overwrite = FALSE, validate = FALSE)
}
\arguments{
\item{model}{a n2kModel object}

\item{base}{the base location to store the model}

\item{project}{will be a relative path within the base location}

\item{overwrite}{should an existing object be overwritten? Defaults to TRUE}

\item{validate}{check that the object is valid before storing it. Defaults to TRUE}
}
\value{
a data.frame with the fingerprint, status and status_fingerprint as stored
}
\description{
store the n2kModel and retrieve its fingerprints
}
