% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_weight.R
\name{calculate_weight}
\alias{calculate_weight}
\title{Calculate the survey weights for the observations}
\usage{
calculate_weight(observation, result.channel)
}
\arguments{
\item{observation}{the dataframe with observations. Should be the output of
\code{\link{read_observation}}}

\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Adds the survey weights and the stratum to the observations
}

