% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_raw}
\alias{connect_raw}
\title{Make a git connection to the raw data}
\usage{
connect_raw(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Make a git connection to the raw data
}

