% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_manifest_objects.R
\name{copy_manifest_objects}
\alias{copy_manifest_objects}
\title{Copy all object of the most recent manifest}
\usage{
copy_manifest_objects(base, project, target_folder = ".", overwrite = FALSE)
}
\arguments{
\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{target_folder}{Where to store the objects}

\item{overwrite}{Should an existing object be overwritten?
Defaults to \code{TRUE}.}
}
\description{
Copy all object of the most recent manifest
}
