% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id}
\alias{datasource_id}
\title{Get the id of the data source}
\usage{
datasource_id(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Get the id of the data source
}

