% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_lincomb.R
\name{get_linear_lincomb}
\alias{get_linear_lincomb}
\title{Calculate the matrix for linear combinations of strata}
\usage{
get_linear_lincomb(
  dataset,
  time_var,
  stratum_var = "stratum",
  stratum_weights,
  formula
)
}
\arguments{
\item{dataset}{the raw dataset.}

\item{time_var}{variable indicating the time.}

\item{stratum_var}{variable containing the stratum information.}

\item{stratum_weights}{a dataframe with stratum weights.}

\item{formula}{the formula for the trend component}
}
\description{
Calculate the matrix for linear combinations of strata
}
