% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonlinear_lincomb.R
\name{get_nonlinear_lincomb}
\alias{get_nonlinear_lincomb}
\title{Calculate the matrix for linear combinations of strata}
\usage{
get_nonlinear_lincomb(
  dataset,
  stratum_weights,
  time_var,
  label_var,
  stratum_var
)
}
\arguments{
\item{dataset}{the raw dataset.}

\item{stratum_weights}{a dataframe with stratum weights.}

\item{time_var}{variable indicating the time.}

\item{label_var}{variable containing the label for every `time_var`.}

\item{stratum_var}{variable containing the stratum information.}
}
\description{
Calculate the matrix for linear combinations of strata
}
