% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonlinear_lincomb.R
\name{get_nonlinear_lincomb}
\alias{get_nonlinear_lincomb}
\title{Calculate the matrix for linear combinations of strata}
\usage{
get_nonlinear_lincomb(dataset, time.var, stratum.var = "fStratum", formula)
}
\arguments{
\item{dataset}{the raw dataset}

\item{time.var}{the name of the time variable}

\item{stratum.var}{the name of the stratum variable}

\item{formula}{the formula for the model.matrix}
}
\description{
Calculate the matrix for linear combinations of strata
}
