% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all datasets and a to do list of models}
\usage{
prepare_analysis(min.observation = 100)
}
\arguments{
\item{min.observation}{The minimum number of positive observations (Count > 0)}
}
\description{
Prepare all datasets and a to do list of models
}

