% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all datasets and a to do list of models}
\usage{
prepare_analysis(repo, base, project, overwrite = FALSE,
  min.observation = 100, min.stratum = 3, min.cycle = 2,
  proportion = 0.15)
}
\arguments{
\item{repo}{a \code{\link[git2rdata]{repository}} object}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to TRUE.}

\item{min.observation}{the required minimum number of occassion where the present was observed, defaults to 100}

\item{min.stratum}{the required minimum number of relevant locations per stratum, defaults to 3}

\item{min.cycle}{the required minimum number of cycles in which the species was present at a location. defaults to 2}

\item{proportion}{minimum proportion of the least important period, defaults to 0.15}
}
\description{
Prepare all datasets and a to do list of models
}
