% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all datasets and a to do list of models}
\usage{
prepare_analysis(
  repo,
  base,
  project,
  overwrite = FALSE,
  seed = 20070315,
  verbose = TRUE,
  min_observation = 100,
  min_stratum = 3,
  min_cycle = 2,
  proportion = 0.15,
  scheme_id = "ABV",
  volume,
  docker = "inbobmk/rn2k:0.7",
  dependencies = c("inbo/n2khelper@v0.4.3", "inbo/n2kanalysis@v0.2.9",
    "inbo/n2kupdate@v0.1.1")
)
}
\arguments{
\item{repo}{a [git2rdata::repository()] object}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to TRUE.}

\item{verbose}{display the progress}

\item{min_stratum}{The required minimum number of relevant locations per
stratum.
Defaults to `3`.}

\item{proportion}{Minimum proportion of the least important period.
Defaults to `0.15`.}

\item{volume}{The argument passed to the '-v' of docker.
Only used when `base` is a local file system.}

\item{docker}{The docker image to use during model fit.}

\item{dependencies}{A vector of R packages as understood by the `repo`
argument of [remotes::install_github()].}
}
\description{
Prepare all datasets and a to do list of models
}
