% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all datasets and a to do list of models}
\usage{
prepare_analysis(raw.connection, analysis.path = ".", min.observation = 100)
}
\arguments{
\item{raw.connection}{a git-connection object to write the output to}

\item{analysis.path}{the path to store the rda files for the analysis}

\item{min.observation}{The minimum number of positive observations
(Count > 0)}
}
\description{
Prepare all datasets and a to do list of models
}

