% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_comparison.R
\name{prepare_analysis_comparison}
\alias{prepare_analysis_comparison}
\title{Prepare comparison among models}
\usage{
prepare_analysis_comparison(dataset, raw.connection, analysis.path)
}
\arguments{
\item{dataset}{A data.frame with the parent analyses}

\item{analysis.path}{the path to store the rda files for the analysis}
}
\description{
Prepare comparison among models
}
