% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_composite.R
\name{prepare_analysis_composite}
\alias{prepare_analysis_composite}
\title{Prepare the analysis files for the composite indices}
\usage{
prepare_analysis_composite(frequency, type, species_group, location_group,
  models, base, project, overwrite = FALSE)
}
\arguments{
\item{frequency}{a indicator of the frequency of the analysis}

\item{type}{the model type}

\item{species_group}{the species_group ID}

\item{location_group}{the location_group ID}

\item{models}{a data.frame with the metadata of the models}

\item{base}{the base location to store the model}

\item{project}{will be a relative path within the base location}

\item{overwrite}{should an existing object be overwritten? Defaults to TRUE}
}
\description{
Prepare the analysis files for the composite indices
}
