% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_composite.R
\name{prepare_analysis_composite}
\alias{prepare_analysis_composite}
\title{Prepare the analysis files for the composite indices}
\usage{
prepare_analysis_composite(dataset, raw.connection, analysis.path)
}
\arguments{
\item{dataset}{the raw dataset}

\item{raw.connection}{a git-connection object to write the output to}

\item{analysis.path}{the path to store the rda files for the analysis}
}
\description{
Prepare the analysis files for the composite indices
}

