% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_analysis_dataset.R
\name{prepare_analysis_dataset}
\alias{prepare_analysis_dataset}
\title{Create the analysis dataset based on the available raw data}
\usage{
prepare_analysis_dataset(rawdata.file, observation, raw.connection,
  analysis.path, min.observation)
}
\arguments{
\item{rawdata.file}{The file with the counts per visit}

\item{observation}{the dataframe with the visits and location group}

\item{raw.connection}{a git-connection object to write the output to}

\item{analysis.path}{the path to store the rda files for the analysis}

\item{min.observation}{The minimum number of positive observations
(Count > 0)}
}
\value{
A data.frame with the species id number of rows in the analysis
   dataset, number of precenses in the analysis datset and SHA-1 of the
   analysis dataset or NULL if not enough data.
}
\description{
This analysis fits an unweighted model but adds the stratum effect. The
   indices per year of the different strata are combined with a linear
   combination into a single index per year.
}

