% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_dataset.R
\name{prepare_analysis_dataset}
\alias{prepare_analysis_dataset}
\title{Create the analysis dataset based on the available raw data}
\usage{
prepare_analysis_dataset(species_group, location_group, metadata,
  observation, repo, base, project, overwrite = FALSE,
  min.observation = 100, min.stratum = 3, min.cycle = 2,
  proportion = 0.15)
}
\arguments{
\item{species_group}{the species_group ID}

\item{location_group}{the location_group ID}

\item{metadata}{a dataframe with the metadata of a single import}

\item{observation}{the dataframe with the visits and location group}

\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{base}{the base location to store the model}

\item{project}{will be a relative path within the base location}

\item{overwrite}{should an existing object be overwritten? Defaults to TRUE}

\item{min.observation}{the required minimum number of occassion where the present was observed, defaults to 100}

\item{min.stratum}{the required minimum number of relevant locations per stratum, defaults to 3}

\item{min.cycle}{the required minimum number of cycles in which the species was present at a location. defaults to 2}

\item{proportion}{minimum proportion of the least important period, defaults to 0.15}
}
\value{
A data.frame with the species id number of rows in the analysis
   dataset, number of precenses in the analysis datset and SHA-1 of the
   analysis dataset or NULL if not enough data.
}
\description{
This analysis fits an unweighted model but adds the stratum effect. The
   indices per year of the different strata are combined with a linear
   combination into a single index per year.
}
