% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_dataset.R
\name{prepare_analysis_dataset}
\alias{prepare_analysis_dataset}
\title{Create the analysis dataset based on the available raw data}
\usage{
prepare_analysis_dataset(species_group, location_group, metadata,
  observation, repo, base, project, overwrite = FALSE,
  min.observation = 100, min.stratum = 3, min.cycle = 2,
  proportion = 0.15)
}
\arguments{
\item{species_group}{the species_group ID}

\item{location_group}{the location_group ID}

\item{metadata}{a dataframe with the metadata of a single import}

\item{observation}{The dataframe with the visits and location group.}

\item{repo}{a \code{\link[git2rdata]{repository}} object}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to TRUE.}

\item{min.observation}{The required minimum number of occassion where the
present was observed.
Defaults to `100`.}

\item{min.stratum}{The required minimum number of relevant locations per
stratum.
Defaults to `3`.}

\item{min.cycle}{The required minimum number of cycles in which the species
was present at a location.
Defaults to `2`.}

\item{proportion}{Minimum proportion of the least important period.
Defaults to `0.15`.}
}
\value{
A data.frame with the species id number of rows in the analysis
   dataset, number of precenses in the analysis datset and SHA-1 of the
   analysis dataset or NULL if not enough data.
}
\description{
This analysis fits an unweighted model but adds the stratum effect. The
   indices per year of the different strata are combined with a linear
   combination into a single index per year.
}
