% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_dataset.R
\name{prepare_analysis_dataset}
\alias{prepare_analysis_dataset}
\title{Create the analysis dataset based on the available raw data}
\usage{
prepare_analysis_dataset(
  scheme_id,
  species_group_id,
  species,
  location_group_id,
  seed = 20070315,
  observation,
  repo,
  base,
  project,
  overwrite = FALSE,
  min_observation = 100,
  min_stratum = 3,
  min_cycle = 2,
  proportion = 0.15,
  verbose = TRUE,
  family
)
}
\arguments{
\item{scheme_id}{A single integer holding the id of the scheme.}

\item{species_group_id}{the species_group ID}

\item{species}{The fingerprint of the species.}

\item{location_group_id}{the location_group ID}

\item{seed}{A single integer used as a seed for all calculations.
A random seed will be inserted when missing.}

\item{observation}{The dataframe with the visits and location group.}

\item{repo}{a [git2rdata::repository()] object}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to TRUE.}

\item{min_observation}{The minimal number of observations for a species.
Defaults to `100`.}

\item{min_stratum}{The required minimum number of relevant locations per
stratum.
Defaults to `3`.}

\item{proportion}{Minimum proportion of the least important period.
Defaults to `0.15`.}

\item{verbose}{display the progress}
}
\value{
A data.frame with the species id number of rows in the analysis
   dataset, number of presences in the analysis dataset and SHA-1 of the
   analysis dataset or NULL if not enough data.
}
\description{
This analysis fits an unweighted model but adds the stratum effect. The
   indices per year of the different strata are combined with a linear
   combination into a single index per year.
}
