% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_analysis_dataset.R
\name{prepare_analysis_dataset}
\alias{prepare_analysis_dataset}
\title{Create the analysis dataset based on the available raw data}
\usage{
prepare_analysis_dataset(rawdata.file, observation, rawdata.path = "abv",
  analysis.path = "analysis", min.observation = 100)
}
\arguments{
\item{rawdata.file}{The file with the counts per visit}

\item{observation}{the dataframe with the visits and location group}

\item{rawdata.path}{the path in the git repository holding the raw data}

\item{analysis.path}{the path to store the rda files for the analysis}

\item{min.observation}{The minimum number of positive observations (Count > 0)}
}
\value{
A data.frame with the species id number of rows in the analysis dataset, number of precenses in the analysis datset and SHA-1 of the analysis dataset or NULL if not enough data.
}
\description{
The analysis dataset is saved to a rda file with the SHA-1 as name.
}

