% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare all datasets
Prepare all datasets and store the version into a git repository}
\usage{
prepare_dataset()
}
\description{
Prepare all datasets
Prepare all datasets and store the version into a git repository
}

