% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare all datasets
Prepare all datasets and store the version into a git repository}
\usage{
prepare_dataset(result.channel, source.channel, attribute.connection,
  raw.connection, scheme.id, verbose = TRUE)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}

\item{source.channel}{An open ODBC connection to the source database}

\item{attribute.connection}{a git-connection object to the attribute data}

\item{raw.connection}{a git-connection object to write the output to}

\item{scheme.id}{The id of the scheme}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Prepare all datasets
Prepare all datasets and store the version into a git repository
}

