% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare all datasets
Prepare all datasets and store the version into a git repository}
\usage{
prepare_dataset(username, password, verbose = TRUE, develop = TRUE)
}
\arguments{
\item{username}{the username for the git repository}

\item{password}{the user's password for the git repository}

\item{verbose}{Display a progress bar when TRUE (default)}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Prepare all datasets
Prepare all datasets and store the version into a git repository
}

