% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_location.R
\name{prepare_dataset_location}
\alias{prepare_dataset_location}
\title{import and store location metadata}
\usage{
prepare_dataset_location(origin, repo, end_date, strict = TRUE)
}
\arguments{
\item{origin}{\code{DBI} connection to the source database}

\item{repo}{a \code{\link[git2rdata:repository]{git2rdata::repository()}} object}

\item{end_date}{the latest date to import.
Default to now.}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}
}
\description{
import and store location metadata
}
