% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations and save them to git and the results database}
\usage{
prepare_dataset_observation(origin, result, repo, end.date, species)
}
\arguments{
\item{origin}{`dbplyr::src_dbi` connection to the source database}

\item{result}{An open ODBC connection to the results database}

\item{repo}{a \code{\link[git2rdata]{repository}} object}

\item{end.date}{the latest date to import. Default to now.}

\item{species}{a data.frame with the NBN key and fingerprint for each species}
}
\description{
Read the observations and save them to git and the results database
}
