% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations and save them to git and the results database}
\usage{
prepare_dataset_observation(source.channel, result.channel, raw.connection,
  attribute.connection)
}
\arguments{
\item{source.channel}{An open ODBC connection to the source database}

\item{result.channel}{An open RODBC connection to the results database}

\item{raw.connection}{a git-connection object to write the output to}

\item{attribute.connection}{a git-connection object to the attribute data}
}
\description{
Read the observations and save them to git and the results database
}

