% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations and save them to git and the results database}
\usage{
prepare_dataset_observation(source.channel, result.channel, raw.git)
}
\arguments{
\item{source.channel}{An open ODBC channel to the source database}

\item{result.channel}{An open RODBC connection to the results database}

\item{raw.git}{A git_connection to store the rawdata}
}
\description{
Read the observations and save them to git and the results database
}

