% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations and save them to git and the results database}
\usage{
prepare_dataset_observation(
  origin,
  repo,
  end_date,
  min_observation = 100,
  strict = TRUE
)
}
\arguments{
\item{origin}{\code{DBI} connection to the source database}

\item{repo}{a \code{\link[git2rdata:repository]{git2rdata::repository()}} object}

\item{end_date}{the latest date to import.
Default to now.}

\item{min_observation}{The minimum number of observations for taking a
species into account.
Defaults to \code{100}.}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}
}
\description{
Read the observations and save them to git and the results database
}
