% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species groups and save them to git and the results database}
\usage{
prepare_dataset_species(origin, result, repo, scheme.id, end.date)
}
\arguments{
\item{origin}{`dbplyr::src_dbi` connection to the source database}

\item{result}{An open ODBC connection to the results database}

\item{repo}{a \code{\link[git2rdata]{repository}} object}

\item{scheme.id}{the fingerprint of the scheme}

\item{end.date}{the latest date to import. Default to now.}
}
\description{
Read the species groups and save them to git and the results database
}
