% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species groups and save them to git and the results database}
\usage{
prepare_dataset_species(origin, repo, end_date, strict = FALSE, db_scheme = "")
}
\arguments{
\item{origin}{\code{DBI} connection to the source database}

\item{repo}{a \code{\link[git2rdata:repository]{git2rdata::repository()}} object}

\item{end_date}{the latest date to import.
Default to now.}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}

\item{db_scheme}{optional database scheme name.}
}
\description{
Read the species groups and save them to git and the results database
}
