% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_observation_species.R
\name{prepare_dataset_species_observation}
\alias{prepare_dataset_species_observation}
\title{Read the counts and save them to git}
\usage{
prepare_dataset_species_observation(this.species, observation, result.channel,
  source.channel, raw.connection, scheme.id, first.year, last.year)
}
\arguments{
\item{this.species}{a dataframe with the ExternalCode and the SpeciesGroupID}

\item{observation}{a data.frame with observations}

\item{result.channel}{An open RODBC connection to the results database}

\item{source.channel}{An open ODBC connection to the source database}

\item{raw.connection}{a git-connection object to write the output to}

\item{scheme.id}{The id of the scheme}

\item{first.year}{oldest year of the data}

\item{last.year}{most recent year of the data}
}
\description{
Read the counts and save them to git
}
