% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_observation.R
\name{read_observation}
\alias{read_observation}
\title{Return the set of all visited locations}
\usage{
read_observation(source.channel, result.channel)
}
\arguments{
\item{source.channel}{An open ODBC connection to the source database}

\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Return the set of all visited locations
}

