% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_observation_species.R
\name{read_observation_species}
\alias{read_observation_species}
\title{Read the observations of a given species}
\usage{
read_observation_species(species.id, source.channel)
}
\arguments{
\item{species.id}{The external id of the species}

\item{source.channel}{An open ODBC connection to the source database}
}
\description{
Read the observations of a given species
}
