% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_observation_species.R
\name{read_observation_species}
\alias{read_observation_species}
\title{Read the observations of a given species}
\usage{
read_observation_species(species.id)
}
\arguments{
\item{species.id}{The id of the species}
}
\description{
Read the observations of a given species
}

