% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_specieslist.R
\name{read_specieslist}
\alias{read_specieslist}
\title{Read the list of species}
\usage{
read_specieslist(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\value{
A list with two components: \code{Species} holds the names of all species, \code{Speciesgroup} lists the species per user defined group.
}
\description{
Read the list of species
}
\examples{
read_specieslist()
}

