% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_relevant.R
\name{select_relevant}
\alias{select_relevant}
\title{Select the relevant observation of a species}
\usage{
select_relevant(observation, species, datafield, repo,
  min.observation = 100, min.stratum = 3, min.cycle = 2,
  proportion = 0.15)
}
\arguments{
\item{observation}{The dataframe with the visits and location group.}

\item{species}{The fingerprint of the species.}

\item{datafield}{A named vector with the datafield id for `"sample"` and
`"observation"`.}

\item{repo}{a \code{\link[git2rdata]{repository}} object}

\item{min.observation}{The required minimum number of occassion where the
present was observed.
Defaults to `100`.}

\item{min.stratum}{The required minimum number of relevant locations per
stratum.
Defaults to `3`.}

\item{min.cycle}{The required minimum number of cycles in which the species
was present at a location.
Defaults to `2`.}

\item{proportion}{Minimum proportion of the least important period.
Defaults to `0.15`.}
}
\description{
Relevant locations have at least two observations in different cycle.
Relevant periods have average numbers of at least 15% of the most important
period.
}
