% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage}
\alias{storage}
\title{store the n2kModel and retrieve its fingerprints}
\usage{
storage(model, base, project, overwrite = FALSE, validate = FALSE)
}
\arguments{
\item{model}{a n2kModel object}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to \code{TRUE}.}

\item{validate}{Check that the object is valid before storing it.
Defaults to \code{TRUE}.}
}
\value{
a data.frame with the fingerprint, status and status_fingerprint as
stored
}
\description{
store the n2kModel and retrieve its fingerprints
}
